+++ otx/algorithms/detection/adapters/mmdet/data/__init__.py
@@ -14,6 +14,7 @@
 # See the License for the specific language governing permissions
 # and limitations under the License.
 
-from .mpa_det_dataset import MPADetDataset
+from .dataset import MPADetDataset
+from .pipelines import LoadAnnotationFromOTXDataset, LoadImageFromOTXDataset
 
-__all__ = ["MPADetDataset"]
+__all__ = ["MPADetDataset", "LoadAnnotationFromOTXDataset", "LoadImageFromOTXDataset"]
