+++ otx/algorithms/detection/adapters/mmdet/data/dataset.py
@@ -23,6 +23,7 @@ from mmdet.datasets.builder import DATASETS
 from mmdet.datasets.custom import CustomDataset
 from mmdet.datasets.pipelines import Compose
 
+from otx.algorithms.common.utils.data_utils import get_old_new_img_indices
 from otx.api.entities.dataset_item import DatasetItemEntity
 from otx.api.entities.datasets import DatasetEntity
 from otx.api.entities.label import Domain, LabelEntity
@@ -236,3 +237,30 @@ class OTXDataset(CustomDataset):
         dataset_item = self.otx_dataset[idx]
         labels = self.labels
         return get_annotation_mmdet_format(dataset_item, labels, self.domain)
+
+
+@DATASETS.register_module()
+class MPADetDataset(OTXDataset):
+    """MPADetDataset for Class-Incremental Learning."""
+
+    def __init__(self, **kwargs):
+        dataset_cfg = kwargs.copy()
+        _ = dataset_cfg.pop("org_type", None)
+        new_classes = dataset_cfg.pop("new_classes", [])
+        super().__init__(**dataset_cfg)
+
+        test_mode = kwargs.get("test_mode", False)
+        if test_mode is False:
+            self.img_indices = get_old_new_img_indices(self.labels, new_classes, self.otx_dataset)
+
+    def get_cat_ids(self, idx):
+        """Get category ids by index.
+
+        Args:
+            idx (int): Index of data.
+
+        Returns:
+            list[int]: All categories in the image of specified index.
+        """
+
+        return self.get_ann_info(idx)["labels"].astype(int).tolist()
