+++ otx/algorithms/detection/adapters/mmdet/utils/config_utils.py
