+++ otx/algorithms/detection/adapters/openvino/__init__.py
@@ -1,0 +1,4 @@
+"""OTX Adapters - openvino."""
+
+# Copyright (C) 2022 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
