+++ otx/algorithms/detection/tasks/nncf.py
@@ -40,7 +40,7 @@ from mpa.utils.config_utils import remove_custom_hook
 from mpa.utils.logger import get_logger
 
 from otx.algorithms.common.adapters.mmcv.hooks import OTXLoggerHook
-from otx.algorithms.detection.adapters.mmdet.config_utils import (
+from otx.algorithms.detection.adapters.mmdet.utils.config_utils import (
     patch_config,
     prepare_for_training,
     remove_from_config,
