+++ otx/cli/datasets/object_detection/dataset.py
@@ -14,8 +14,7 @@
 # See the License for the specific language governing permissions
 # and limitations under the License.
 
-from detection_tasks.extension.datasets.data_utils import load_dataset_items_coco_format
-
+from otx.algorithms.detection.utils.data_utils import load_dataset_items_coco_format
 from otx.api.entities.datasets import DatasetEntity
 from otx.api.entities.label import Domain
 from otx.api.entities.subset import Subset
